almost finish subshell_parse but has memleaks
